pyproject.toml
[tool.poetry]
name = "sep-2023-drf"
version = "0.1.0"
description = ""
authors = ["Vitalii Demchyshyn <demchyshyn.v87@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.15.1"
mysqlclient = "^2.2.4"


[tool.poetry.group.dev.dependencies]
django-dotenv = "^1.4.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
